move deps to root
