--- src/deps/index.ts
@@ -1,2 +1,0 @@
-export * from './dependencies';
-export * from './model';
