+++ src/java/projenrc.ts
@@ -2,7 +2,7 @@ import { dirname, join } from 'path';
 import { existsSync, mkdirpSync, writeFileSync } from 'fs-extra';
 import { PROJEN_VERSION } from '../common';
 import { Component } from '../component';
-import { DependencyType } from '../deps';
+import { DependencyType } from '../dependencies';
 import { readJsiiManifest } from '../inventory';
 import { Project } from '../project';
 import { Pom } from './pom';
